autoscaling/requirements.txt
#Installation instructions: First run pip install setuptools-rust and then pip install -r requirements.txt
#Create a serice principal by running az ad sp create-for-rbac --name localtest-sp-rbac (WITHOUT --skip-assignment)
adal==1.2.7
azure-common==1.1.27
azure-core==1.13.0
azure-identity==1.5.0
azure-mgmt-compute==20.0.0
azure-mgmt-core==1.2.2
azure-mgmt-monitor==2.0.0
azure-mgmt-network==18.0.0
azure-mgmt-resource==16.1.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
daemons==1.3.1
google-auth==1.30.0
idna==2.10
isodate==0.6.0
kubernetes==12.0.1
msal==1.11.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.0
pkg-resources==0.0.0
portalocker==1.7.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
